+++ app/src/main/java/com/durand/dogedex/ui/user_fragment/list_can_report_lost/CanReportLostAdapter.kt
@@ -1,4 +1,4 @@
-package com.durand.dogedex.ui.user_fragment.can_report_lost
+package com.durand.dogedex.ui.user_fragment.list_can_report_lost
 
 import android.view.LayoutInflater
 import android.view.View
