+++ app/src/main/java/com/durand/dogedex/ui/user_fragment/list_can_report_lost/CanReportLostFragment.kt
@@ -1,8 +1,7 @@
-package com.durand.dogedex.ui.user_fragment.can_report_lost
+package com.durand.dogedex.ui.user_fragment.list_can_report_lost
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
