+++ app/src/main/java/com/durand/dogedex/ui/user_fragment/list_can_report_lost/CanReportLostViewModel.kt
@@ -1,6 +1,4 @@
-@file:Suppress("SpellCheckingInspection")
-
-package com.durand.dogedex.ui.user_fragment.my_can_register
+package com.durand.dogedex.ui.user_fragment.list_can_report_lost
 
 import android.util.Log
 import androidx.lifecycle.LiveData
@@ -9,23 +7,27 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.durand.dogedex.data.ApiResponseStatus
 import com.durand.dogedex.data.repository.NewOficialRepository
-import com.durand.dogedex.data.response.oficial.ListarCanResponse
+import com.durand.dogedex.data.response.oficial.ListarCanPerdidoResponse
 import kotlinx.coroutines.launch
 
-class MyCanRegisterViewModel(
+class CanReportLostViewModel(
     private val repository: NewOficialRepository = NewOficialRepository()
 ) : ViewModel() {
 
-    private val _list = MutableLiveData<List<ListarCanResponse>>(emptyList())
-    val list: LiveData<List<ListarCanResponse>> = _list
+    private val _list = MutableLiveData<List<ListarCanPerdidoResponse>>()
+    val list: LiveData<List<ListarCanPerdidoResponse>> = _list
 
     private val _isLoading = MutableLiveData(false)
     val isLoading: LiveData<Boolean> = _isLoading
 
+    init {
+        listar()
+    }
+
     fun listar() = viewModelScope.launch {
         _isLoading.postValue(true)
         try {
-            when (val res: ApiResponseStatus<List<ListarCanResponse>> = repository.listarMascota()) {
+            when (val res: ApiResponseStatus<List<ListarCanPerdidoResponse>> = repository.listarMascotaPerdida()) {
                 is ApiResponseStatus.Error -> {
                     Log.d("josue", "Login Error")
                     _isLoading.postValue(false)
@@ -47,4 +49,5 @@ class MyCanRegisterViewModel(
             _isLoading.postValue(false)
         }
     }
+
 }
