+++ app/src/main/java/com/durand/dogedex/ui/user_fragment/list_my_can_register/MyCanRegisterFragment.kt
@@ -1,7 +1,6 @@
-package com.durand.dogedex.ui.user_fragment.my_can_register
+package com.durand.dogedex.ui.user_fragment.list_my_can_register
 
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -28,8 +27,7 @@ class MyCanRegisterFragment : Fragment() {
 
         viewModel.listar()
         viewModel.list.observe(viewLifecycleOwner) {
-            binding.canReportLostRecyclerView.layoutManager =
-                LinearLayoutManager(requireContext()) // o `binding.root.context`
+            binding.canReportLostRecyclerView.layoutManager = LinearLayoutManager(requireContext()) // o `binding.root.context`
             binding.canReportLostRecyclerView.adapter = MyCanRegisterAdapter(it)
         }
