+++ app/src/main/res/layout/fragment_can_report_lost.xml
@@ -4,7 +4,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        tools:context=".ui.user_fragment.can_report_lost.CanReportLostFragment">
+        tools:context=".ui.user_fragment.list_can_report_lost.CanReportLostFragment">
 
     <TextView
             android:id="@+id/text"
